+++ api/src/controllers/orderControllers/getOrderByUserIdController.js
@@ -11,7 +11,7 @@ const getOrderByUserIdController = async (userId) =>{
           include: Food,
         },
       });
-      console.log("nada", cart.dataValues);
+      // console.log("nada", cart.dataValues);
       return cart.dataValues
 }
